rework progressbar
